+++ src/blocks/carousel-slide/save.js
@@ -1,0 +1,31 @@
+/**
+ * WordPress dependencies
+ */
+import { useBlockProps, InnerBlocks } from '@wordpress/block-editor';
+
+/**
+ * Save component for the carousel slide block.
+ *
+ * @param {Object} props            Block props.
+ * @param {Object} props.attributes Block attributes.
+ * @return {Element} Element to render.
+ */
+export default function save( { attributes } ) {
+	const className = [
+		'swiper-slide',
+		attributes.verticalAlign &&
+			`are-vertically-aligned-${ attributes.verticalAlign }`,
+	]
+		.filter( Boolean )
+		.join( ' ' );
+
+	const blockProps = useBlockProps.save( {
+		className,
+	} );
+
+	return (
+		<div { ...blockProps }>
+			<InnerBlocks.Content />
+		</div>
+	);
+}